RtsUtils: Use `size_t` instead of `int` where appropriate
Functions like `stgMallocBytes` take a size parameter which was of type
`int`, but is commonly used as `stgMallocBytes (sizeof (...))`. This is
problematic because the `sizeof` operator returns `size_t` so that on 64
bit systems, in this common use case the `size_t` parameter would be
truncated to 32 bits when passed to `stgMallocBytes` where it was cast
back to `size_t`.

Test Plan: Validate on Linux, OS X and Windows

Reviewers: austin, hvr, bgamari, simonmar, hsyl20

Reviewed By: hvr, bgamari, simonmar, hsyl20

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D2201

+++ rts/RtsUtils.h
@@ -18,12 +18,12 @@
 void initAllocator(void);
 void shutdownAllocator(void);
 
-void *stgMallocBytes(int n, char *msg)
+void *stgMallocBytes(size_t n, char *msg)
     GNUC3_ATTRIBUTE(__malloc__);
 
-void *stgReallocBytes(void *p, int n, char *msg);
+void *stgReallocBytes(void *p, size_t n, char *msg);
 
-void *stgCallocBytes(int n, int m, char *msg)
+void *stgCallocBytes(size_t n, size_t m, char *msg)
      GNUC3_ATTRIBUTE(__malloc__);
 
 char *stgStrndup(const char *s, size_t n);
